requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
django-redis==5.4.0
celery==5.4.0
redis==5.0.7
cryptography==43.0.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
pillow==11.3.0